Update component Square, add confetti on win
