+++ package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "canvas-confetti": "1.6.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
